Replace table layout with grid layout
